make style as sub-screen for karaoke skin editor.
