--- osu.Game.Rulesets.Karaoke.Tests/Screens/TestSceneStyle.cs
@@ -1,44 +1,0 @@
-﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
-// See the LICENCE file in the repository root for full licence text.
-
-using NUnit.Framework;
-using osu.Framework.Allocation;
-using osu.Game.Rulesets.Edit;
-using osu.Game.Rulesets.Karaoke.Beatmaps;
-using osu.Game.Rulesets.Karaoke.Screens.Skin.Style;
-using osu.Game.Rulesets.Karaoke.Skinning;
-using osu.Game.Rulesets.Karaoke.Tests.Beatmaps;
-using osu.Game.Rulesets.Karaoke.Tests.Editor;
-using osu.Game.Screens.Edit;
-using osu.Game.Skinning;
-
-namespace osu.Game.Rulesets.Karaoke.Tests.Screens
-{
-    [TestFixture]
-    public class TestSceneStyle : EditorSubScreenTestScene<StyleScreen>
-    {
-        protected override StyleScreen CreateEditor() => new();
-
-        [Cached(typeof(EditorBeatmap))]
-        [Cached(typeof(IBeatSnapProvider))]
-        private readonly EditorBeatmap editorBeatmap;
-
-        public TestSceneStyle()
-        {
-            var beatmap = new TestKaraokeBeatmap(null);
-            var karaokeBeatmap = new KaraokeBeatmapConverter(beatmap, new KaraokeRuleset()).Convert() as KaraokeBeatmap;
-            editorBeatmap = new EditorBeatmap(karaokeBeatmap);
-        }
-
-        [BackgroundDependencyLoader]
-        private void load(SkinManager skinManager)
-        {
-            Beatmap.Value = CreateWorkingBeatmap(editorBeatmap.PlayableBeatmap);
-
-            // because skin is compared by id, so should change id to let skin manager info knows that skin has been changed.
-            var defaultSkin = DefaultKaraokeSkin.Default;
-            defaultSkin.ID = 100;
-            skinManager.CurrentSkinInfo.Value = defaultSkin;
-        }
-    }
-}
